+++ src/v1/MediaType/index.ts
@@ -35,3 +35,4 @@
 export * from "./isMediaType";
 export * from "./mustBeMediaType";
 export * from "./MediaTypeParts";
+export * from "./parseMediaType";